Add files via upload
This is a completed project. This is an android application used by Flight Paramedics to search through their medical protocols and view them quickly. This uses a pdf document and supports internal pdf hyperlinking. The main PDF document is left out of this for continuity reasons. The PDF would just need to be placed into the assets folder with the name of "smog.pdf".